# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "VK TPA Website API Documentation",
        "title": "TPA-Website",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/hotel/add_hotel_room": {
            "post": {
                "description": "Add a new hotel room",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "Add hotel room",
                "parameters": [
                    {
                        "description": "Hotel Room",
                        "name": "hotelRoom",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hotel room created successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/room/add_hotel_room_picture": {
            "post": {
                "description": "Add a new room picture",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Room"
                ],
                "summary": "Add room picture",
                "parameters": [
                    {
                        "description": "Room Picture",
                        "name": "roomPicture",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room picture created successfully!",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}